+++ sdk/dotnet/src/Consts.fs
@@ -37,9 +37,9 @@ type EventType =
 
 module Consts =
     let [<Literal>] DatastarKey               = "datastar"
-    let [<Literal>] Version                   = "0.21.0-beta1"
-    let [<Literal>] VersionClientByteSize     = 33222
-    let [<Literal>] VersionClientByteSizeGzip = 12210
+    let [<Literal>] Version                   = "0.21.0-beta2"
+    let [<Literal>] VersionClientByteSize     = 33232
+    let [<Literal>] VersionClientByteSizeGzip = 12213
 
     /// Default: TimeSpan.FromMilliseconds 300
     let DefaultSettleDuration = TimeSpan.FromMilliseconds 300