+++ sdk/go/consts.go
@@ -6,9 +6,9 @@ import "time"
 
 const (
     DatastarKey = "datastar"
-    Version                   = "0.21.0-beta1"
-    VersionClientByteSize     = 33222
-    VersionClientByteSizeGzip = 12210
+    Version                   = "0.21.0-beta2"
+    VersionClientByteSize     = 33232
+    VersionClientByteSizeGzip = 12213
 
     //region Default durations