+++ sdk/php/src/Consts.php
@@ -10,9 +10,9 @@ use starfederation\datastar\enums\FragmentMergeMode;
 class Consts
 {
     public const DATASTAR_KEY = 'datastar';
-    public const VERSION = '0.21.0-beta1';
-    public const VERSION_CLIENT_BYTE_SIZE = 33222;
-    public const VERSION_CLIENT_BYTE_SIZE_GZIP = 12210;
+    public const VERSION = '0.21.0-beta2';
+    public const VERSION_CLIENT_BYTE_SIZE = 33232;
+    public const VERSION_CLIENT_BYTE_SIZE_GZIP = 12213;
 
     // The default duration for settling during merges. Allows for CSS transitions to complete.
     public const DEFAULT_SETTLE_DURATION = 300;