+++ sdk/dotnet/src/Consts.fs
@@ -37,7 +37,7 @@ type EventType =
 
 module Consts =
     let [<Literal>] DatastarKey               = "datastar"
-    let [<Literal>] Version                   = "0.21.0-beta2"
+    let [<Literal>] Version                   = "0.21.0"
     let [<Literal>] VersionClientByteSize     = 33602
     let [<Literal>] VersionClientByteSizeGzip = 12354