+++ sdk/go/consts.go
@@ -6,7 +6,7 @@ import "time"
 
 const (
     DatastarKey = "datastar"
-    Version                   = "0.21.0-beta2"
+    Version                   = "0.21.0"
     VersionClientByteSize     = 33602
     VersionClientByteSizeGzip = 12354