+++ sdk/php/src/Consts.php
@@ -10,7 +10,7 @@ use starfederation\datastar\enums\FragmentMergeMode;
 class Consts
 {
     public const DATASTAR_KEY = 'datastar';
-    public const VERSION = '0.21.0-beta2';
+    public const VERSION = '0.21.0';
     public const VERSION_CLIENT_BYTE_SIZE = 33602;
     public const VERSION_CLIENT_BYTE_SIZE_GZIP = 12354;